+++ windirstat/Pages/PageAdvanced.cpp
@@ -45,7 +45,7 @@ void CPageAdvanced::DoDataExchange(CDataExchange* pDX)
     DDX_Text(pDX, IDC_FOLDER_HISTORY_COUNT, m_folderHistoryCount);
     DDX_CBIndex(pDX, IDC_COMBO_THREADS, m_scanningThreads);
     DDX_CBIndex(pDX, IDC_HASH_ALGORITHM, m_fileHashAlgorithm);
-    DDX_CBIndex(pDX, IDC_SCAN_IO_PRIORITY, m_scanIoPriority);
+    DDX_CBIndex(pDX, IDC_PROCESS_PRIORITY, m_processPriority);
 }
 
 BEGIN_MESSAGE_MAP(CPageAdvanced, CMFCPropertyPage)
@@ -54,7 +54,7 @@ BEGIN_MESSAGE_MAP(CPageAdvanced, CMFCPropertyPage)
     ON_BN_CLICKED(IDC_EXCLUDE_PROTECTED_DIRECTORY, OnSettingChanged)
     ON_CBN_SELENDOK(IDC_COMBO_THREADS, OnSettingChanged)
     ON_CBN_SELENDOK(IDC_HASH_ALGORITHM, OnSettingChanged)
-    ON_CBN_SELENDOK(IDC_SCAN_IO_PRIORITY, OnSettingChanged)
+    ON_CBN_SELENDOK(IDC_PROCESS_PRIORITY, OnSettingChanged)
     ON_BN_CLICKED(IDC_EXCLUDE_VOLUME_MOUNT_POINTS, OnSettingChanged)
     ON_BN_CLICKED(IDC_EXCLUDE_JUNCTIONS, OnSettingChanged)
     ON_BN_CLICKED(IDC_EXCLUDE_SYMLINKS_DIRECTORY, OnSettingChanged)
@@ -97,15 +97,14 @@ BOOL CPageAdvanced::OnInitDialog()
     m_processHardlinks = COptions::ProcessHardlinks;
     m_scanningThreads = COptions::ScanningThreads - 1;
     m_fileHashAlgorithm = COptions::FileHashAlgorithm;
-    m_scanIoPriority = COptions::ScanIoPriority;
+    m_processPriority = COptions::ProcessPriority;
     m_largestFileCount = std::to_wstring(COptions::LargeFileCount.Obj()).c_str();
     m_folderHistoryCount = std::to_wstring(COptions::FolderHistoryCount.Obj()).c_str();
 
-    if (auto* pIoPrio = static_cast<CComboBox*>(GetDlgItem(IDC_SCAN_IO_PRIORITY)); pIoPrio != nullptr)
+    if (auto* priorityCombo = static_cast<CComboBox*>(GetDlgItem(IDC_PROCESS_PRIORITY)); priorityCombo != nullptr)
     {
-        pIoPrio->AddString(Localization::Lookup(IDS_IO_PRIORITY_NORMAL).c_str());
-        pIoPrio->AddString(Localization::Lookup(IDS_IO_PRIORITY_BELOW_NORMAL).c_str());
-        pIoPrio->AddString(Localization::Lookup(IDS_IO_PRIORITY_IDLE).c_str());
+        for (const auto& priority : SplitString(Localization::Lookup(IDS_PRIORITY_LEVELS), L','))
+            priorityCombo->AddString(priority.c_str());
     }
 
     UpdateData(FALSE);
@@ -141,7 +140,8 @@ void CPageAdvanced::OnOK()
     COptions::ProcessHardlinks = (FALSE != m_processHardlinks);
     COptions::ScanningThreads = m_scanningThreads + 1;
     COptions::FileHashAlgorithm = m_fileHashAlgorithm;
-    COptions::ScanIoPriority = m_scanIoPriority;
+    COptions::ProcessPriority = m_processPriority;
+    SetProcessPriority(COptions::ProcessPriority);
     COptions::LargeFileCount = std::stoi(m_largestFileCount.GetString());
     COptions::FolderHistoryCount = std::stoi(m_folderHistoryCount.GetString());