+++ windirstat/WinDirStat.cpp
@@ -366,9 +366,6 @@ BOOL CDirStatApp::InitInstance()
     // Prevent state saving
     m_bSaveState = FALSE;
 
-    // Set process I/O priority to high for better disk scanning performance
-    SetProcessIoPriorityHigh();
-
     // Load default language just to get bootstrapped
     Localization::LoadResource(MAKELANGID(LANG_ENGLISH, SUBLANG_NEUTRAL));
 
@@ -376,6 +373,7 @@ BOOL CDirStatApp::InitInstance()
     SetPortableMode(true, true);
 
     COptions::LoadAppSettings();
+    SetProcessPriority(COptions::ProcessPriority);
     LoadStdProfileSettings(0);
 
     // Silently restart elevated conditionally before any expensive initialization