+++ CHANGELOG.md
@@ -5,6 +5,7 @@ Enhancements
 - Added flame graph as an alternative visualization view (thanks @P403n1x87)
 - Added additional window layouts
 - Added per-user Explorer context menu registration
+- Added configurable process CPU and I/O priority
 - Added File Watcher toolbar controls
 - Added cancellable file hashing progress (thanks @harryytm)
 - Added folder frames color customization and draw threshold (thanks @harryytm)